add command parameter for Galaxy parser
Due to this line https://github.com/hexylena/argparse2tool/blob/cca59571ad18f0bcc7b863ebbdd42a3480801aaf/argparse/__init__.py#L144 this allows to get the xml only for a subcommand. No idea if this was intended like this? For consistency this could be added to the cwl part as well.